+++ src/api/controllers/user.controller.js
@@ -1,29 +1,18 @@
 const service = require('../services/user.service')
 const User = require('../models/User')
-const { checkCPF } = require('../../helpers/cpfChecker')
-
-
+const { checkSchema } = require('../../helpers/schemaChecker')
 
 const newAccount = async (request, h) => {
     try {
         const user = new User(request.payload)
-        if (checkCPF(user.getCpf())) {
-            const payloadKeys = Object.entries(user)
 
-            payloadKeys.forEach((entry) => {
-                const value = entry[1]
-
-                if (!value) {
-                    throw new Error(`${entry[0]} obrigatório.`)
-                }
-            })
-
+        if (checkSchema(user)) {
             const result = await service.createAccount(user)
             return h.response(result).code(201)
-        } else {
-            return h.response({ message: "Cpf inválido" }).code(400)
         }
+
     } catch (err) {
+        console.log(err)
         return h.response({ error: err.message }).code(err.statusCode)
     }
 }
